@did-btcr2/bitcoin@0.2.0 + @did-btcr2/method@0.15.0 - fixes coupling bitcoin to node packages
