+++ packages/bitcoin/src/bitcoin.ts
@@ -1,13 +1,15 @@
 import { MethodError } from '@did-btcr2/common';
 import { networks } from 'bitcoinjs-lib';
-import { DEFAULT_REST_CLIENT_CONFIG } from './constants.js';
+import { BitcoinRestClient } from './client/rest/index.js';
+import { BitcoinCoreRpcClient } from './client/rpc/index.js';
 import { getNetwork } from './network.js';
-import { BitcoinRest } from './rest-client.js';
 import { AvailableNetworks, BitcoinClientConfig } from './types.js';
+import { DEFAULT_BITCOIN_NETWORK_CONFIG } from './constants.js';
 
 export type BitcoinNetworkConfig = {
   name: keyof AvailableNetworks;
-  rest: BitcoinRest;
+  rpc?: BitcoinCoreRpcClient;
+  rest: BitcoinRestClient;
   config: BitcoinClientConfig;
   data: networks.Network;
 };
@@ -17,14 +19,15 @@ export type BitcoinNetworkConfigMap = {
 };
 
 /**
- * General class to house the Bitcoin client connections, client config and various utility methods.
- * @name Bitcoin
- * @type {Bitcoin}
+ * The BitcoinNetworkConnection class contains relevant data and methods for interacting with different Bitcoin network.
+ * @name BitcoinNetworkConnection
+ * @type {BitcoinNetworkConnection}
  */
-export class Bitcoin {
+export class BitcoinNetworkConnection {
   public network: BitcoinNetworkConfig;
   public mainnet?: BitcoinNetworkConfig;
-  public testnet?: BitcoinNetworkConfig;
+  public testnet3?: BitcoinNetworkConfig;
+  public testnet4?: BitcoinNetworkConfig;
   public signet?: BitcoinNetworkConfig;
   public mutinynet?: BitcoinNetworkConfig;
   public regtest?: BitcoinNetworkConfig;
@@ -36,11 +39,7 @@ export class Bitcoin {
    * @throws {MethodError} If no configs is passed and BITCOIN_NETWORK_CONFIG is missing or invalid.
    */
   constructor(configs?: BitcoinNetworkConfigMap) {
-    const BITCOIN_NETWORK_CONFIG = process.env.BITCOIN_NETWORK_CONFIG ?? JSON.stringify(configs ?? {
-      regtest : {
-        rest : DEFAULT_REST_CLIENT_CONFIG
-      }
-    });
+    const BITCOIN_NETWORK_CONFIG = process.env.BITCOIN_NETWORK_CONFIG ?? JSON.stringify(configs ?? DEFAULT_BITCOIN_NETWORK_CONFIG);
 
     if(!BITCOIN_NETWORK_CONFIG) {
       throw new MethodError(
@@ -63,7 +62,7 @@ export class Bitcoin {
     const networkConfigs: Record<string, BitcoinClientConfig> = JSON.parse(BITCOIN_NETWORK_CONFIG);
 
     // Set a list of available networks
-    const networks: (keyof AvailableNetworks)[] = ['mainnet', 'testnet', 'signet', 'mutinynet', 'regtest'];
+    const networks: (keyof AvailableNetworks)[] = ['mainnet', 'testnet3', 'testnet4', 'signet', 'mutinynet', 'regtest'];
 
     // Iterate over the networks and create the client connections
     for (const network of networks) {
@@ -72,7 +71,8 @@ export class Bitcoin {
         this[network] = {
           name   : network,
           config : networkConfig,
-          rest   : new BitcoinRest(networkConfig.rest ?? DEFAULT_REST_CLIENT_CONFIG) ,
+          rpc    : new BitcoinCoreRpcClient(networkConfig.rpc),
+          rest   : new BitcoinRestClient(networkConfig.rest) ,
           data   : getNetwork(network),
         };
       }
@@ -96,6 +96,38 @@ export class Bitcoin {
   }
 
   /**
+   * Get the Bitcoin network configuration for a specific network.
+   * @param {keyof AvailableNetworks} network - The Bitcoin network (e.g., 'mainnet', 'testnet3', 'signet', 'regtest').
+   * @returns {Bitcoin} The Bitcoin object.
+   */
+  public getNetworkConnection(network: string): BitcoinNetworkConnection {
+    const availableNetwork = network as keyof AvailableNetworks;
+    if (!this[availableNetwork]) {
+      throw new MethodError(
+        `No configuration found for network='${availableNetwork}'`,
+        'MISSING_CONFIG_FOR_NETWORK'
+      );
+    }
+    return this;
+  }
+
+  /**
+   * Sets the active Bitcoin network.
+   * @param {keyof AvailableNetworks} active - The Bitcoin network to set as active (e.g., 'mainnet', 'testnet3', 'signet', 'regtest').
+   * @throws {MethodError} If no configuration is found for the specified network.
+   */
+  public setActiveNetwork(active: string): void {
+    const availableNetwork = active as keyof AvailableNetworks;
+    if (!this[availableNetwork]) {
+      throw new MethodError(
+        `No configuration found for network='${availableNetwork}'`,
+        'MISSING_CONFIG_FOR_NETWORK'
+      );
+    }
+    this.network = this[availableNetwork] as BitcoinNetworkConfig;
+  }
+
+  /**
    * Converts Bitcoin (BTC) to satoshis (SAT).
    * @param {number} btc - The amount in BTC.
    * @returns {number} The amount in SAT.
@@ -114,4 +146,4 @@ export class Bitcoin {
   };
 }
 
-export const bitcoin = new Bitcoin();
+export const bitcoin = new BitcoinNetworkConnection();
