--- packages/bitcoin/src/index.node.ts
@@ -1,9 +1,0 @@
-export * from './bitcoin.node.js';
-export * from './constants.js';
-export * from './errors.js';
-export * from './interface.js';
-export * from './network.js';
-export * from './rest-client.js';
-export * from './rpc-client.js';
-export * from './taproot.js';
-export * from './types.js';
