+++ packages/bitcoin/src/index.ts
@@ -1,0 +1,14 @@
+export * from './bitcoin.js';
+export * from './client/rest/address.js';
+export * from './client/rest/block.js';
+export * from './client/rest/index.js';
+export * from './client/rest/transaction.js';
+export * from './client/rpc/index.js';
+export * from './client/rpc/interface.js';
+export * from './client/rpc/json-rpc.js';
+export * from './client/utils.js';
+export * from './constants.js';
+export * from './errors.js';
+export * from './network.js';
+export * from './taproot.js';
+export * from './types.js';
