+++ packages/bitcoin/tests/bitcoin.spec.ts
@@ -1,0 +1,12 @@
+import { expect } from 'chai';
+import { BitcoinNetworkConnection } from '../src/bitcoin.js';
+
+/**
+ * Bitcoin Network Connection - Browser Tests
+ */
+describe('Bitcoin Network Connection Tests', () => {
+  it('should initialize the browser Bitcoin class', () => {
+    const bitcoin = new BitcoinNetworkConnection();
+    expect(bitcoin).to.exist;
+  });
+});
