+++ packages/method/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@did-btcr2/method",
-    "version": "0.14.6",
+    "version": "0.15.0",
     "type": "module",
     "description": "Javascript/TypeScript reference implementation of did:btcr2 method, a censorship resistant DID Method using the Bitcoin blockchain as a Verifiable Data Registry to announce changes to the DID document. Core package of the did-btcr2-js monorepo.",
     "main": "./dist/cjs/index.js",
@@ -94,6 +94,7 @@
         "do:lib:btcr2": "${PWD}/lib/do-lib.sh btcr2"
     },
     "dependencies": {
+        "@bitcoinerlab/secp256k1": "^1.2.0",
         "@did-btcr2/bitcoin": "workspace:*",
         "@did-btcr2/common": "workspace:*",
         "@did-btcr2/cryptosuite": "workspace:*",
@@ -109,7 +110,6 @@
         "@web5/common": "^1.1.0",
         "@web5/crypto": "^1.0.6",
         "@web5/dids": "^1.2.0",
-        "bitcoin-core": "^5.0.0",
         "bitcoinjs-lib": "7.0.0-rc.0",
         "canonicalize": "^2.1.0",
         "dotenv": "^16.5.0",
